app/api/requirements.txt
flask
flask-cors
TTS
sounddevice
torchaudio
torch
numpy
python-dotenv
librosa
soundfile
scipy
pymongo
gunicorn
pydub